Work around `ip -o monitor address' inconsistency.
Sometimes, non-deterministically, the interface name is numbered
as in 'ifN', at least for resign events and inet family, even if
the name is known since it is printed at the end of the line. Thus,
in that case, we take the last name instead. When monitoring with
two instances, one might show the name, the other not. Further, brd
check is no more hardcoded to lo, instead is assigned when available.